♻️ Create separate handler file for CreateClientCommand
Moved the CreateClientCommandHandler logic to its own file to improve organization and maintainability. This change separates the command definition from its handler, adhering to the single responsibility principle.
