+++ Application/Features/Clients/Commands/CreateClientCommand/CreateClientCommand.cs
@@ -1,7 +1,4 @@
-﻿using Application.Interfaces;
-using Application.Wrappers;
-using AutoMapper;
-using Domain.Entities;
+﻿using Application.Wrappers;
 using MediatR;
 
 namespace Application.Features.Clients.Commands.CreateClientCommand
@@ -14,26 +11,5 @@ namespace Application.Features.Clients.Commands.CreateClientCommand
         public string Telefono { get; set; }
         public string Email { get; set; }
         public string Direccion { get; set; }
-    }
-
-    public class CreateClientCommandHandler : IRequestHandler<CreateClientCommand, Response<int>>
-    {
-        // Ayuda a mapear automaticamente
-        private readonly IRepositoryAsync<Client> _repositoryAsync;
-        private readonly IMapper _mapper;
-
-        public CreateClientCommandHandler(IRepositoryAsync<Client> repositoryAsync, IMapper mapper)
-        {
-            _repositoryAsync = repositoryAsync;
-            _mapper = mapper;
-        }
-
-        public async Task<Response<int>> Handle(CreateClientCommand request, CancellationToken cancellationToken)
-        {
-            var newRecord = _mapper.Map<Client>(request);
-            var data = await _repositoryAsync.AddAsync(newRecord);
-
-            return new Response<int>(data.Id);
-        }
     }
 }
